requirements.txt
#openai
pillow
azure-storage-blob 
selenium
requests
#opencv-python
#python-dotenv
#pytesseract